--- src/components/CityCard.vue
@@ -1,33 +1,0 @@
-<template>
-    <div></div>
-  </template>
-  
-  <script setup>
-  import axios from "axios";
-  import { useRoute } from "vue-router";
-  const route = useRoute();
-  const getWeatherData = async () => {
-    try {
-      const weatherData = await axios.get(
-        `https://api.openweathermap.org/data/2.5/onecall?lat=${route.query.
-            lat}&lon=${route.query.lng}&exclude={part}&
-            appid=a4330711038a63eef7ecdaaac48623e1&units=imperial`
-      );
-      // cal current date & time
-      const localOffset = new Date().getTimezoneOffset() * 60000;
-      const utc = weatherData.data.current.dt * 1000 + localOffset;
-      weatherData.data.currentTime =
-        utc + 1000 * weatherData.data.timezone_offset;
-      // cal hourly weather offset
-      weatherData.data.hourly.forEach((hour) => {
-        const utc = hour.dt * 1000 + localOffset;
-        hour.currentTime =
-          utc + 1000 * weatherData.data.timezone_offset;
-      });
-      return weatherData.data;
-    } catch (err) {
-      console.log(err);
-    }
-  };
-  const weatherData = await getWeatherData();
-  </script>
